apply review changes
